+++ drv_system.h
@@ -66,10 +66,10 @@
 #endif
 
 void systemInit(void);
-void delayMicroseconds(uint32_t us);
+void delayMicroseconds(uint64_t us);
 void delay(uint32_t ms);
 
-uint32_t micros(void);
+uint64_t micros(void);
 uint32_t millis(void);
 
 // Backup SRAM R/W
